mcpscreen/pyproject.toml
[project]
name = "mcpscreen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.4",
    "mss>=10.0.0",
    "pillow>=11.2.1",
]
